Create README.md
ARENA INNOVATION - Hans Marius Schuster is „2020 Mission Innovation Champion“ representing Germany within the Innovation Challenge IC8: Renewable and Clean Hydrogen (http://www.mission-innovation.net/2020/09/02/champions-corner-hans-marius-schuster-to-link-up-efforts-for-hydrogen-society-platform-2/) and the founder of ARENA INNOVATION, which propagates the energy and digital transformation for the Fuel Cell & Hydrogen (FCH) domain by means of Artificial Intelligence (AI) technologies under his “H2 Metaverse” digital platform. The “H2Metaverse” is a global innovative AI-powered cloud-based digital platform under which a networked system of a variety of decentralized individual platforms for hydrogen and fuel cell technologies is created, all following a common standard – the Gaia-X standard. It is based on the next generation of Gaia-X data infrastructure for an open, transparent, and secure digital ecosystem where data and services can be made available, collected, and shared in a trusted environment. For a first practical pilot, the "Digital Twin" (DT) smart service for the development and simulation of a self sustaining hydrogen economy and production of hydrogen electrolyzers and fuel cells is standardized and made available to many cloud service providers. Other smart services, such as e-ID (digital/electronic identity), "Guarantee of Origin for Green Hydrogen" (see EU project CertifHy) or so-called digital "Smart Contracts for Difference (Cfd's)" (see the German national funding instrument H2Global) also belong to this concept.

Mr. Schuster is an IT engineer from the University of Applied Sciences Giessen-Friedberg, Germany and a certified European Research and Innovation Funding Programmes Professional from the School of International Business and Entrepreneurship - Steinbeis University Berlin. He has built up consortia for and advised many large-scale projects for European programs like H2020, FCH-JU and CEF. He represents the standardization committee ASRO/CT 395 - Hydrogen in Energy Systems and is an active member of the GAIA-X Working Group for defining use cases for the energy and mobility domain within the Workstream User Ecosystems and Requirements.

Mr. Schuster’s goal is to establish a so-called "H2 Metaverse" to foster the energy transition and true international collaboration beyond technical, political, and geographical borders. As a Mission Innovation Champion, the impact of his vision will definitely increase significantly.